+++ packages/firestore/src/model/document.ts
@@ -16,13 +16,14 @@
  */
 
 import { SnapshotVersion } from '../core/snapshot_version';
-import { fail } from '../util/assert';
+import { assert, fail } from '../util/assert';
 
 import { DocumentKey } from './document_key';
 import { FieldValue, JsonObject, ObjectValue } from './field_value';
 import { FieldPath } from './path';
 
 import * as api from '../protos/firestore_proto_api';
+import * as obj from '../util/obj';
 
 export interface DocumentOptions {
   hasLocalMutations?: boolean;
@@ -59,16 +60,19 @@ export class Document extends MaybeDocument {
   readonly hasLocalMutations: boolean;
   readonly hasCommittedMutations: boolean;
 
+  /**
+   * A cache of canonicalized FieldPaths to FieldValues that have already been
+   * deserialized in `getField()`.
+   */
+  private fieldValueCache?: Map<string, FieldValue | null>;
+
   constructor(
     key: DocumentKey,
     version: SnapshotVersion,
-    readonly data: ObjectValue,
     options: DocumentOptions,
-    /**
-     * Memoized serialized form of the document for optimization purposes (avoids repeated
-     * serialization). Might be undefined.
-     */
-    readonly proto?: api.Document
+    private objectValue?: ObjectValue,
+    readonly proto?: api.Document,
+    private readonly converter?: (value: api.Value) => FieldValue
   ) {
     super(key, version);
     this.hasLocalMutations = !!options.hasLocalMutations;
@@ -76,7 +80,51 @@ export class Document extends MaybeDocument {
   }
 
   field(path: FieldPath): FieldValue | null {
-    return this.data.field(path);
+    if (this.objectValue) {
+      return this.objectValue.field(path);
+    } else {
+      assert(
+        this.proto !== undefined && this.converter !== undefined,
+        'Expected proto and converter to be defined.'
+      );
+
+      if (!this.fieldValueCache) {
+        // TODO(b/136090445): Remove the cache when `getField` is no longer
+        // called during Query ordering.
+        this.fieldValueCache = new Map<string, FieldValue>();
+      }
+
+      const canonicalPath = path.canonicalString();
+
+      let fieldValue = this.fieldValueCache.get(canonicalPath);
+
+      if (fieldValue === undefined) {
+        // Instead of deserializing the full Document proto, we only
+        // deserialize the value at the requested field path. This speeds up
+        // Query execution as query filters can discard documents based on a
+        // single field.
+        let protoValue: api.Value | undefined = this.proto!.fields[
+          path.firstSegment()
+        ];
+        for (let i = 1; protoValue !== undefined && i < path.length; ++i) {
+          if (protoValue.mapValue) {
+            protoValue = protoValue.mapValue.fields[path.get(i)];
+          } else {
+            protoValue = undefined;
+          }
+        }
+
+        if (protoValue === undefined) {
+          fieldValue = null;
+        } else {
+          fieldValue = this.converter!(protoValue);
+        }
+
+        this.fieldValueCache.set(canonicalPath, fieldValue);
+      }
+
+      return fieldValue!;
+    }
   }
 
   fieldValue(path: FieldPath): unknown {
@@ -84,8 +132,29 @@ export class Document extends MaybeDocument {
     return field ? field.value() : undefined;
   }
 
+  data(): ObjectValue {
+    if (!this.objectValue) {
+      assert(
+        this.proto !== undefined && this.converter !== undefined,
+        'Expected proto and converter to be defined.'
+      );
+
+      let result = ObjectValue.EMPTY;
+      obj.forEach(this.proto!.fields, (key: string, value: api.Value) => {
+        result = result.set(new FieldPath([key]), this.converter!(value));
+      });
+      this.objectValue = result;
+
+      // Once objectValue is computed, values inside the fieldValueCache are no
+      // longer accessed.
+      this.fieldValueCache = undefined;
+    }
+
+    return this.objectValue;
+  }
+
   value(): JsonObject<unknown> {
-    return this.data.value();
+    return this.data().value();
   }
 
   isEqual(other: MaybeDocument | null | undefined): boolean {
@@ -93,9 +162,9 @@ export class Document extends MaybeDocument {
       other instanceof Document &&
       this.key.isEqual(other.key) &&
       this.version.isEqual(other.version) &&
-      this.data.isEqual(other.data) &&
       this.hasLocalMutations === other.hasLocalMutations &&
-      this.hasCommittedMutations === other.hasCommittedMutations
+      this.hasCommittedMutations === other.hasCommittedMutations &&
+      this.data().isEqual(other.data())
     );
   }