+++ packages/firestore/src/model/mutation.ts
@@ -353,9 +353,14 @@ export class SetMutation extends Mutation {
     // have held.
 
     const version = mutationResult.version;
-    return new Document(this.key, version, this.value, {
-      hasCommittedMutations: true
-    });
+    return new Document(
+      this.key,
+      version,
+      {
+        hasCommittedMutations: true
+      },
+      this.value
+    );
   }
 
   applyToLocalView(
@@ -370,9 +375,14 @@ export class SetMutation extends Mutation {
     }
 
     const version = Mutation.getPostMutationVersion(maybeDoc);
-    return new Document(this.key, version, this.value, {
-      hasLocalMutations: true
-    });
+    return new Document(
+      this.key,
+      version,
+      {
+        hasLocalMutations: true
+      },
+      this.value
+    );
   }
 
   extractBaseValue(maybeDoc: MaybeDocument | null): null {
@@ -434,9 +444,14 @@ export class PatchMutation extends Mutation {
     }
 
     const newData = this.patchDocument(maybeDoc);
-    return new Document(this.key, mutationResult.version, newData, {
-      hasCommittedMutations: true
-    });
+    return new Document(
+      this.key,
+      mutationResult.version,
+      {
+        hasCommittedMutations: true
+      },
+      newData
+    );
   }
 
   applyToLocalView(
@@ -452,9 +467,14 @@ export class PatchMutation extends Mutation {
 
     const version = Mutation.getPostMutationVersion(maybeDoc);
     const newData = this.patchDocument(maybeDoc);
-    return new Document(this.key, version, newData, {
-      hasLocalMutations: true
-    });
+    return new Document(
+      this.key,
+      version,
+      {
+        hasLocalMutations: true
+      },
+      newData
+    );
   }
 
   extractBaseValue(maybeDoc: MaybeDocument | null): null {
@@ -478,7 +498,7 @@ export class PatchMutation extends Mutation {
   private patchDocument(maybeDoc: MaybeDocument | null): ObjectValue {
     let data: ObjectValue;
     if (maybeDoc instanceof Document) {
-      data = maybeDoc.data;
+      data = maybeDoc.data();
     } else {
       data = ObjectValue.EMPTY;
     }
@@ -550,10 +570,15 @@ export class TransformMutation extends Mutation {
     );
 
     const version = mutationResult.version;
-    const newData = this.transformObject(doc.data, transformResults);
-    return new Document(this.key, version, newData, {
-      hasCommittedMutations: true
-    });
+    const newData = this.transformObject(doc.data(), transformResults);
+    return new Document(
+      this.key,
+      version,
+      {
+        hasCommittedMutations: true
+      },
+      newData
+    );
   }
 
   applyToLocalView(
@@ -573,10 +598,15 @@ export class TransformMutation extends Mutation {
       maybeDoc,
       baseDoc
     );
-    const newData = this.transformObject(doc.data, transformResults);
-    return new Document(this.key, doc.version, newData, {
-      hasLocalMutations: true
-    });
+    const newData = this.transformObject(doc.data(), transformResults);
+    return new Document(
+      this.key,
+      doc.version,
+      {
+        hasLocalMutations: true
+      },
+      newData
+    );
   }
 
   extractBaseValue(maybeDoc: MaybeDocument | null): ObjectValue | null {